Logger: done

+++ Wms.Web/src/Business/Concrete/BoxService.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using Serilog;
 using Wms.Web.Business.Abstract;
 using Wms.Web.Business.Dto;
 using Wms.Web.Business.Extensions;
@@ -6,7 +7,6 @@ using Wms.Web.Common.Exceptions;
 using Wms.Web.Repositories.Interfaces;
 using Wms.Web.Store.Entities.Concrete;
 using Wms.Web.Store.Entities.Specifications;
-using static Serilog.ILogger;
 
 namespace Wms.Web.Business.Concrete;
 
@@ -58,7 +58,11 @@ internal sealed class BoxService : IBoxService
                     cancellationToken: cancellationToken);
         }
         
-        return _mapper.Map<IReadOnlyCollection<BoxDto>>(entities);
+        var boxDto = _mapper.Map<IReadOnlyCollection<BoxDto>>(entities);
+        
+        Log.Logger.Information("Got {Count} boxes", boxDto.Count);
+
+        return boxDto;
     }
 
     /// <inheritdoc />
@@ -70,15 +74,24 @@ internal sealed class BoxService : IBoxService
         
         if (box != null)
         {
+            Log.Logger.Error("Box already exist: id={Id}", boxDto.Id);
             throw new EntityAlreadyExistException(boxDto.Id);
         }
-        
+
         var palette = await _paletteRepository
-                          .GetByIdAsync(boxDto.PaletteId, cancellationToken)
-                      ?? throw new EntityNotFoundException(boxDto.PaletteId);
+            .GetByIdAsync(boxDto.PaletteId, cancellationToken);
+        
+        if (palette == null)
+        {
+            Log.Logger.Error("Palette not found: id={Id}", boxDto.PaletteId);
+            
+            throw new EntityNotFoundException(boxDto.PaletteId);
+        }
 
         if (palette.DeletedAt is not null)
         {
+            Log.Logger.Error("Palette was deleted: id={Id}", boxDto.PaletteId);
+            
             throw new EntityWasDeletedException(palette.Id);
         }
 
@@ -91,6 +104,8 @@ internal sealed class BoxService : IBoxService
         await _boxRepository.CreateAsync(box, cancellationToken);
 
         await _paletteService.RefreshAsync(box.PaletteId, cancellationToken);
+        
+        Log.Logger.Information("Box created: \n{Box}", box.ToString());
     }
 
     /// <inheritdoc />
@@ -109,21 +124,36 @@ internal sealed class BoxService : IBoxService
         BoxDto boxDto, 
         CancellationToken cancellationToken)
     {
-        var box = await _boxRepository.GetByIdAsync(boxDto.Id, cancellationToken)
-                  ?? throw new EntityNotFoundException(boxDto.Id);
-
+        var box = await _boxRepository.GetByIdAsync(boxDto.Id, cancellationToken);
+        
+        if (box == null)
+        {
+            Log.Logger.Error("Box wasn't found. ID={Id}", boxDto.Id);
+            throw new EntityNotFoundException(boxDto.Id);
+        }
+        
         if (box.DeletedAt is not null)
         {
+            Log.Logger.Error("Box was deleted. ID={BoxId}", box.Id);
+            
             throw new EntityWasDeletedException(box.Id);
         }
 
         var oldPalette = box.PaletteId;
         
-        var palette = await _paletteRepository.GetByIdAsync(boxDto.PaletteId, cancellationToken)
-                      ?? throw new EntityNotFoundException(boxDto.PaletteId);
+        var palette = await _paletteRepository.GetByIdAsync(boxDto.PaletteId, cancellationToken);
+            
+        if (palette == null)
+        {
+            Log.Logger.Error("Palette wasn't found. ID={PaletteId}", boxDto.PaletteId);
+                
+            throw new EntityNotFoundException(boxDto.PaletteId);
+        }
         
         if (palette.DeletedAt is not null)
         {
+            Log.Logger.Error("Palette was deleted. ID={PaletteId}", palette.Id);
+            
             throw new EntityWasDeletedException(palette.Id);
         }
         
@@ -138,6 +168,8 @@ internal sealed class BoxService : IBoxService
         await _paletteService.RefreshAsync(palette.Id, cancellationToken);
 
         await _paletteService.RefreshAsync(oldPalette, cancellationToken);
+        
+        Log.Logger.Information("Box updated: \n {Box}", box.ToString());
     }
 
     /// <inheritdoc />
@@ -145,16 +177,25 @@ internal sealed class BoxService : IBoxService
         Guid id, 
         CancellationToken cancellationToken)
     {
-        var box = await _boxRepository.GetByIdAsync(id, cancellationToken)
-            ?? throw new EntityNotFoundException(id);
+        var box = await _boxRepository.GetByIdAsync(id, cancellationToken);
+        
+        if (box == null)
+        {
+            Log.Logger.Error("Box wasn't found. ID={Id}", id);
+            throw new EntityNotFoundException(id);
+        }
         
         if (box.DeletedAt is not null)
         {
+            Log.Logger.Warning("Box id={Id} already deleted. Return", id);
+            
             return;
         }
 
         await _boxRepository.DeleteAsync(id, cancellationToken);
 
         await _paletteService.RefreshAsync(box.PaletteId, cancellationToken);
+        
+        Log.Logger.Information("Box id={Id} deleted", id);
     }
 }
